feat: refactoring architecture

+++ src/app/kenzie/system.py
@@ -1,19 +1,26 @@
+import os.path
 from json import dump
-import os
+from pathlib import Path
 
 
-def mk_database():
-    print("Creating database...")
+def create_database(path: str):
+    """
+    Run first migration
+    :param path: Base path to start the database
+    :return: Nothing
+    """
+    full_path = path + "/database.json"
 
     empty_json_data = {"data": []}
-    base_path: str = os.environ.get("DATABASE_PATH")
 
-    # Creating folder
-    os.mkdir(base_path)
+    file_not_exists = not os.path.exists(full_path)
 
-    # Creating json file
-    full_path = base_path + "/database.json"
-    with open(full_path, "w+") as json_file:
-        dump(empty_json_data, json_file, indent=4)
+    if file_not_exists:
+        print("Creating database...")
+        # Generating nested directories
+        Path(path).mkdir(parents=True, exist_ok=True)
 
-    print(f"Database created at {full_path}...")
+        # Creating json file
+        with open(full_path, "w+") as json_file:
+            dump(empty_json_data, json_file, indent=4)
+        print(f"Database created at {full_path}...")
